Rewrote GameManager, resolve #13
