+++ PuzzleGame/Views/StartGameDialog.cs
@@ -101,6 +101,5 @@ public class StartGameDialog : Form
         };
         IsTimeGame = _gameTypeTimeRadioButton.Checked;
         DialogResult = DialogResult.OK;
-        Close();
     }
 }
